almost finished refactoring, need to work on making 'wrong' and 'correct' buttons appear when answer is revealed
